+++ test/test_fusion_adaptive_timeout.ml
@@ -24,7 +24,6 @@ let sample_judge model : Fusion_policy.judge_spec =
   ; jlabel = ""
   ; jsystem_prompt = "judge"
   ; jweb_tools = false
-  ; jmax_tool_calls = 0
   ; jmax_output_tokens = None
   ; jtimeout_s = 1.0
   ; jmax_timeout_s = None